only push image if new

+++ sh/dockerpush.sh
@@ -27,16 +27,16 @@ fi
 dpush() {
     IMAGE="${IMAGE_BASE}-$1:$2"
     URL="${DOCKER_LOGIN_SERVER}/${IMAGE}"
-    # if ! docker manifest inspect "${URL}" &> /dev/null ; then
-    echo "pushing ${IMAGE} to ${URL}"
+    if ! docker pull "${URL}" &> /dev/null ; then
+        echo "pushing ${IMAGE} to ${URL}"
 
-    docker tag "${IMAGE}" "${URL}"
-    docker push "${URL}"
-    docker rmi "${URL}"
-    # else
-    #     echo "${URL} already exists"
-    # fi
-    sleep 1
+        docker tag "${IMAGE}" "${URL}"
+        docker push "${URL}"
+        docker rmi "${URL}"
+        sleep 1
+    else
+        echo "${URL} already exists"
+    fi
 }
 
 dpush "worker" "${IMAGE_TAG}"
@@ -52,12 +52,12 @@ docker pull --platform linux/amd64 "${QDRANT_BASE}"
 
 IMAGE_QDRANT="${IMAGE_BASE}-qdrant:${QDRANT_DOCKER_VERSION}"
 URL_QDRANT="${DOCKER_LOGIN_SERVER}/${IMAGE_QDRANT}"
-# if ! docker manifest inspect "${URL_QDRANT}" &> /dev/null ; then
-echo "pushing ${QDRANT_BASE} to ${URL_QDRANT}"
+if ! docker pull "${URL_QDRANT}" &> /dev/null ; then
+    echo "pushing ${QDRANT_BASE} to ${URL_QDRANT}"
 
-docker tag "${QDRANT_BASE}" "${URL_QDRANT}"
-docker push "${URL_QDRANT}"
-docker rmi "${URL_QDRANT}"
-# else
-#     echo "${URL_QDRANT} already exists"
-# fi
+    docker tag "${QDRANT_BASE}" "${URL_QDRANT}"
+    docker push "${URL_QDRANT}"
+    docker rmi "${URL_QDRANT}"
+else
+    echo "${URL_QDRANT} already exists"
+fi
